Added database, init basic api
Set up the Prisma database, and initialized the api. As of this commit, on render a call is made to the api returning the objects in the database, giving some sample data. In Third.js, that data is retrieved asynchronously, and saved to multiple stateful variables. The site kept crashing every time I tried to update the dictionary type object. I forgot how to React in the time I put down this project. Obviously, the data architecture for the call is some stateful dictionary, housing both featured projects. I might have to reconsider exactly how I want this data when I integrate it with the larger Project archive. For that I might have to research databases more too. Anyways, this was just to get the new changes committed. Next on the tree is to try with useEffect (which I think is the right hook). Hopefully updating from there will be easier