fixed clean database call!
Wow that worked. useEffect was the way to go. I can easily access the project data from the database directly within the component. Next commit should focus on cleaning the actual API so I can map out the format of the calls I'll take, as well as cleaning some of the code and catching errors