fix origin url for fetch
This is an attempt to fix fetching data from the api on different systems. One of the errors previously returned before the last commit was that the fetch url was relative instead of absolute. window.location.origin hopefully fixes this issue, and prevents me from having to change out this line in between pushes