implemented api
API is basically finished, might need some tweaking but for the most part it's there. The main issue with this commit was implementing the logic to handle a non full featured project database. useEffect was actually causing an infinite loop. Not including the dependency array didn't solve it, including it with the characterObject that held the response data wasn't it, it was actually including an EMPTY DEP ARRAY. Took me a bit, but I'm still proud that I found that. Now there is no loop, and the switch case should parse data to the stateful object approriately. Next commit should replace the lorem ipsum text I had previously with the default data / data from the database