completed database for projects
Finally, the featured database portion is complete and working. I cleaned up most of the code, so I'm ready to deploy, and merge back with staging, THEN deploy to prod. Next on the list should probably be populating it with real data, as well as changing the default photo dynamically. Afterwards, I'll probably revisit the contact page thing.