Completed Contact Page
Completed the Contact Page, added an svg logo with a neat animation to compliment it. The site looks good and performs well in mobile as well (for this page). After merging back to staging, then prod, next will be the projects page. I still plan to use Supabase for it