loaded data from db, added styling
data is called from db, added the feathers from featured project section and turned them into components, so more refactoring needs to happen to other files in the codebase using these new components for simplicity. The ProjectEntry has link validation as well, which will also remove unnecessary featers in the case of a null or empty link. Next commit should have the tech list for each project, but this time as a component. That will have to be referenced elsewhere in the codebase as well